move inneroutermapspec
